requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
Django==5.1.7
django-cors-headers==3.14.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
django-cors-headers==3.14.0
drf-yasg==1.21.10
pytest==7.2.1
pytest-django==4.5.2
